+++ iati/tests/test_codelists.py
@@ -205,6 +205,14 @@ class TestCodelistEquality(object):
         assert cmp_func_different(codelist, codelist_copy)
 
     @pytest.mark.parametrize('codelist', iati.default.codelists('2.02').values())
+    def test_codelist_diff_completeness_not_equal(self, codelist, cmp_func_different):
+        """Check that two Codelist with the same codes but different completeness are not deemed to be equal."""
+        codelist_copy = copy.deepcopy(codelist)
+        codelist_copy.complete = not codelist.complete
+
+        assert cmp_func_different(codelist, codelist_copy)
+
+    @pytest.mark.parametrize('codelist', iati.default.codelists('2.02').values())
     def test_codelist_diff_num_codes_not_equal(self, codelist, cmp_func_different):
         """Check that two Codelist with the same name but different codes are not deemed to be equal."""
         codelist_copy = copy.deepcopy(codelist)